+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ----------------------------------------
 
+## [1.9.1] - 2026-04-08 [LIVE-LOG-RECONNECT-FIX]
+
+### Added
+- None.
+
+### Changed
+- None.
+
+### Deprecated
+- None.
+
+### Removed
+- None.
+
+### Fixed
+- Live log disconnect detection.
+
+### Security
+- None.
+
+----------------------------------------
+
 ## [1.9.0] - 2026-04-07 [FIND]
 
 ### Added